Code/raspberry/requirements.txt
# Django==4.0.1
autopep8
tqdm
Pillow
torch==1.11.0
torchvision==0.12.0
numpy==1.22.3